+++ PharmaHub.API/Dtos/Sale/SaleDto.cs
@@ -14,8 +14,8 @@ public class SaleMedicamentDto : BaseDto<SaleMedicamentDto, SaleMedicament>
 {
     public int MedicamentId { get; set; }
     public int Quantity { get; set; }
-    public int PPV { get; set; }
+    public float PPV { get; set; }
     public float Discount { get; set; }
     public double TotalPrice { get; set; }
-    public int TVA { get; set; }
+    public float TVA { get; set; }
 }
