+++ PharmaHub.API/Models/SaleMedicament.cs
@@ -5,7 +5,7 @@ public class SaleMedicament : BaseModel
     public int Quantity { get; set; }
     public double PPV { get; set; }
     public double TotalPrice { get; set; }
-    public int TVA { get; set; }
+    public float TVA { get; set; }
     public float Discount { get; set; }
     public int SaleId { get; set; }
     public Sale Sale { get; set; }
